Split of remaining files Updated CMakeLists.txt file to include make -j 2
